Cut LocoNet into Backend and Dispatcher (parser). Introduced Bus for routng.
