+++ platformio.ini
@@ -12,5 +12,7 @@
 platform = https://github.com/platformio/platform-espressif32.git
 board = esp32dev
 framework = arduino
-lib_deps= EEPROM
+lib_deps= 
+    EEPROM
+    Embedded Template Library
 build_flags = -DDEBUG_OUTPUT=true
