+++ src/LocoNetBus.h
@@ -1,0 +1,46 @@
+/**
+ * Takes LocoNet messages from several sources and resends them to 
+ * many destinations. Used for resending messages between LocoNet, 
+ * Loconet over USB-Serial and LbServer.
+ * */
+
+#pragma once
+
+#include <etl/vector.h>
+
+#include "ln_opc.h"
+
+template < class Msg, class Ret >
+class Consumer {
+public:
+    virtual Ret onMessage(const Msg& msg) = 0;
+};
+
+template <class Msg, class Ret, Ret okVal, const size_t MAX_CONSUMERS>
+class Bus {
+public:
+    using MsgConsumer = Consumer<Msg, Ret>;
+
+    Ret broadcast(const Msg &msg, MsgConsumer* sender = nullptr) {
+        Ret ret = okVal;
+        for(const auto & c: consumers) {
+            if(c!=sender) {
+                Ret v = c->onMessage(msg);
+                if(v!=okVal) ret = v;
+            }
+        }
+        return ret;
+    }
+
+    void addConsumer(MsgConsumer * c) {
+        consumers.push_back(c);
+    }
+
+    void removeConsumer(MsgConsumer * c) {
+        consumers.erase( std::remove(consumers.begin(), consumers.end(), c), consumers.end() );
+    }
+
+private:
+    etl::vector<MsgConsumer*, MAX_CONSUMERS> consumers;
+};
+
